+++ packages/devtools_extensions/lib/src/template/_simulated_devtools_environment/_simulated_devtools_controller.dart
@@ -33,17 +33,13 @@ class SimulatedDevToolsController extends DisposableController
   }
 
   void _handleMessage(Event e) {
-    if (e.isMessageEvent) {
-      final messageData = (e as MessageEvent).data!;
-      final messageData = ((e as MessageEvent).data as JSObject).dartify()!;
-      final extensionEvent = DevToolsExtensionEvent.tryParse(messageData);
-      if (extensionEvent != null) {
-        // Do not handle messages that come from the
-        // [_SimulatedDevToolsController] itself.
-        if (extensionEvent.source == '$SimulatedDevToolsController') return;
+    final extensionEvent = tryParseExtensionEvent(e);
+    if (extensionEvent != null) {
+      // Do not handle messages that come from the
+      // [_SimulatedDevToolsController] itself.
+      if (extensionEvent.source == '$SimulatedDevToolsController') return;
 
-        onEventReceived(extensionEvent);
-      }
+      onEventReceived(extensionEvent);
     }
   }