+++ packages/devtools_extensions/CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.0.11
 * Migrate from `dart:html` to `package:web`.
+* Add `utils.dart` library with helper for message event parsing.
 
 ## 0.0.10
 * Bump minimum Dart SDK version to `3.3.0-91.0.dev` and minimum Flutter SDK version to `3.17.0-0.0.pre`.